web/requirements.txt
Django==3.0.6
gunicorn>=20.0.4
psycopg2-binary>=2.8.5
Pillow>=7.1.2
django-environ==0.4.5
django-ckeditor==5.9.0